system_template: fixed the issue of crashing subfunctions.
The subfunctions were crashing when using nested subfunction paths due to a change to the system template which switched the working director of file execution (mcfunction file execution) to the path of the filter in order to enable use of the pathlib.Path. Now the FUNCTION_PATH variable has been remapped to the same path 'BP/functions' but absolute.