modular_mc: Removed concurrent evaluation of MapTs items.
It turns out that they're fundamentally sequential. Previous code was written under incorrect assumption that it's safe to execute
MapTs items concurently when they use onConflict 'stop', but that's not true, because in order to detect the conflict they have to
know that nothing tries to write to the same target. Close #32